todo-list-backend/requirements.txt
Flask
flask-mysql
python-dotenv
cryptography